[FIRRTLToHW] Generalize lowerRegConnect to support HW wires
Generalize the `lowerRegConnect` function such that it also works on
dynamic array indices and can connect to `hw::WireOp` in addition to the
`seq::FirRegOp`. Rename the function to `lowerConnect` to reflect its
expanded scope.

This removes the need to carry a separate `regMapping` for the backedges
used as the `next` value of registers. Instead, `lowerConnect` directly
inspects the operation on the LHS of the connect and does the right
thing if it's a `seq::FirRegOp` or `hw::WireOp`.

The new connect lowering is implemented as follows:

- First, `lowerConnect` walks the unlowered LHS backwards through any
  subindex, subfield, and subaccess ops, to arrive at the root FIRRTL
  wire or reg being updated.
- We then get the lowered value of that wire or reg, which is going to
  be a `seq::FirRegOp` or `hw::WireOp`, and take its current "next" or
  "input" operand. This is essentially the currently assigned value,
  reflecting all the connects we have seen so far.
- We then create new lowerings for any subaccesses on the LHS, on this
  currently-assigned value. This is necessary since the lowered
  subaccesses that were already created earlier access the lowered
  *wire* or *reg*, and not the *current value* assigned to the wire or
  reg.
- Once we have the current value of the subfield we are assigning, we
  take the lowered RHS of the connect and walk the subaccesses
  backwards, splicing the updated value into the subaccessed aggregates,
  until we finally arrive at the root again and have a new RHS for the
  entire wire or reg with the connected value spliced in at the right
  location.
- Finally we update the root `seq::FirRegOp` or `hw::WireOp` such that
  its "next" or "input" operand corresponds to this new RHS value.

In a nutshell, `lowerConnect` materializes connects directly in the
operand of the connect-to wire or reg, instead of creating something
like `sv.assign` at the point of the connect operation.

The lowering of FIRRTL connect and strictconnect now first attempt to
materialize connects in wires and regs directly through `lowerConnect`,
and only if that's not possible (e.g. because the LHS is not a
`seq::FirRegOp` or `hw::WireOp`) does it resort to generating assign
ops.