[FIRRTLToHW] Lower FIRRTL wires to HW wires
Lower `firrtl.wire` ops to `hw.wire` instead of `sv.wire`. The change is
fairly mechanical: instead of creating an `sv.wire` and setting it as
the FIRRTL wire's lowering, we create an `hw.wire` with an initial Z
value as the lowering. Subsequent connects then pick up the change in
`lowerConnect` and replace the Z with updated values accordingly.
Subaccesses are also fully supported. The Z value prevails if the
connects don't fully assign the entire wire.

The bulk of changes in the tests are just replacing `sv.wire` and the
corresponding `sv.assign` with just the `hw.wire`. A few new tests also
specifically check for subaccess connects to wires.