[FIRRTLToHW] Lower memories to seq.firmem
Make `LowerToHW` emit `seq.firmem` operations for memories instead of
emitting generator ops. These memory ops are then lowered to generator
ops later in the firtool pipeline through the Seq `LowerFirMem` pass,
which is already in place.

The main change is replacing the generator op creation in `LowerToHW`
with code that constructs the FirMem and associated ports, which is
conceptually simpler.

This commit also outlines memory lowering tests into a separate
`lower-to-hw-memories.mlir` file and combines it with a few more test
cases.

The only change to firtool Verilog output should be a loss of the
`_combMem` suffix on memories.